Update flow to configure Flink Kafka bootstrap servers (#982)
## Summary
While testing the launching of Flink kafka jobs, I noticed that the way
we're configuring the event stream (kafka) via the FlinkJob & command
line isn't in-line with how we're doing so for Pubsub (which is either
online args / topic info). I've updated the FlinkJob code to not expect
the Kafka bootstrap servers as a separate arg. Users can configure it
using the standard rails we're providing for PubSub and others.

Made some updates to the Flink kafka event driver code as I noticed it
was broken after a few changes that have gone in since it was last
updated.

## Checklist
- [ ] Added Unit Tests
- [X] Covered by existing CI
- [X] Integration tested
- [ ] Documentation update



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

* **New Features**
* Added support for specifying a parent job ID in Flink Kafka job
arguments.
* Enhanced job submission metadata with new properties, including
metadata name, Zipline version, and unique job IDs.

* **Configuration Updates**
* Updated environment variables and Kafka bootstrap server
configurations in various test and join setups.
  * Modified Kafka topic URIs to explicitly specify a custom serializer.
* Swapped resource allocation settings between "backfill" and "upload"
modes in team metadata configuration.

* **Bug Fixes**
* Corrected the property key for Flink checkpoint URI in job submission
properties.

* **Refactor**
* Simplified Flink job configuration by removing redundant Kafka
bootstrap server extraction and usage.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->