Use partitioned BQ table for Kafka / PubSub canary GBs (#991)
## Summary
With the new viz on errors while running batch jobs, I saw that the
kafka / pubsub GBUs were failing as the test data BQ table which I was
using as a source wasn't partitioned. I ETLed a copy that is partitioned
and updated the GBs to use it.

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [X] Integration tested
- [ ] Documentation update



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

* **Chores**
* Updated configuration to use a partitioned event source table for item
event group by and join operations.
* Modified Airflow dependencies and dataset paths to align with the new
partitioned data source.
* Adjusted environment settings for team metadata tests, including
swapping and updating resource allocations for "backfill" and "upload"
modes.
* Reduced the aggregation window for purchase metrics from 7 days to 1
day across relevant group by and join configurations.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->