Switch GBU job to write out data in a partitioned manner (#993)
## Summary
We recently switched to querying partitions to persist table partition
info during batch job runs. This results in the GBU jobs failing as they
don't write out partitioned data and they use the 'ds' field to
partition (and not the user configured table utils partition col). This
PR extends the GBU jobs to write out data partitioned by ds as well as
fixing the Iceberg table format to pass through the partition column
that is passed as part of retrieving partitions (currently it is
discarded and tableUtils.partitionColumn is what is used).

The alternative approach here is to swap the GBU & upload to kv jobs to
use the user configured partition column as part of the writes. Chose
the current approach as it seemed simpler and these GBU tables are
internal to Chronon so we don't need to adhere to a user partitioning
scheme.

Was able to get a successful set of runs with this change in our canary
(gbu -> upload to kv -> streaming).

## Checklist
- [ ] Added Unit Tests
- [X] Covered by existing CI
- [X] Integration tested
- [ ] Documentation update



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

* **Bug Fixes**
* Ensured upload operations use daily partitioning, improving
consistency in data partitioning and table metadata.
* Adjusted job latency metrics for more accurate reporting and added log
statements for job duration.

* **Tests**
* Enhanced tests to verify that upload nodes correctly use daily
partition specifications.

* **Refactor**
* Updated internal methods to improve clarity and correctness in
partition handling logic.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->