chore:clean up staging query (#999)
## Summary

Simplify the stagingquery so we can test a controlled example of
multi-node DAG.

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update


<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":""}
```
-->


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

* **Chores**
* Updated configuration for environment modes, swapping and adjusting
resource settings for "upload" and "backfill" modes.
* Refined metadata in sample staging queries by removing specific
time-of-day and setup parameters.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

Co-authored-by: thomaschow <thomaschow369@gmail.com>