feat: reduce window size for canary (#1005)
## Summary

- Reduce window size so we don't need to generate so much dummy data to
do a successful run.

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update


<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

* **New Features**
* Added new configuration files for GCP training set and staging query,
enabling more flexible data processing and scheduling options.

* **Enhancements**
* Updated aggregation windows for purchase metrics from 3, 14, and 30
days to 1, 3, and 7 days across AWS and GCP pipelines, providing
finer-grained and more recent data insights.
* Introduced new 1-day and 7-day metrics for sum, count, and average
purchase price in various data pipelines and test configurations.

* **Chores**
* Revised environment and cluster configurations for improved resource
management and execution in different modes.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

Co-authored-by: thomaschow <thomaschow369@gmail.com>