ZIP-845 Add sync API to the CLI (#1019)
## Summary

Running a backfill for the same conf again is not resulting in any step
runs.

This is due to in our user flow:

We call the diff API that checks which confs we have locally that differ
in remote db

We call the upload API on the ones that the diff has returned

Call workflow start API

After user 1 on their own branch has done this the first time, if a
second user on a different branch attempts to run with the same confs,
we skip step 2 since we get back an empty diff.

And that's bad because we need to always insert entries to BranchToNode
(which occurs in step 2's upload API)

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

* **New Features**
* Introduced a synchronization step after uploading configurations for
improved consistency.
* Added new request and response types to support synchronization in the
API.

* **Bug Fixes**
* Enhanced error handling during synchronization API calls for increased
reliability.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->