Exclude logback deps on the Flink side (#1143)
## Summary
With the switch to mill, we ended up pulling in a few logback deps by
default into our various targets. This seems to be fine for the service
& spark but hasn't played well with Flink. Our Dataproc flink jobs are
started up by default pointing to a log4j2 properties and if logback if
in the classpath our logging ends up being a bit messy and verbose (we
log at debug level and log to stdout instead of log files). This PR
breaks our logging deps into the api (which our pure lib modules like
api and aggregator depend on) as well as the logback one (which spark &
service depend on). Flink depends just on the api logging dep and ends
up using the default Flink cluster log4j2 logging props and configs.

## Testing
* Kicked off an adhoc pipeline and confirmed that Spark log continue to
be populated, Flink logs are now at INFO+ and being logged to
jobmanager.log & taskmanager.log
* Service logging continues to work as before (via logback)

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [X] Integration tested
- [ ] Documentation update



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- New Features
- Allows choosing your own logging backend by depending on the logging
API only.
- Refactor
- Split logging dependencies to expose the API separately and updated
modules to use the API-only set.
- Chores
- Aligned dependencies across modules to reduce transitive logging
implementations and avoid conflicts.
- Notes
- No functional changes; existing behavior remains the same, with leaner
default dependencies.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->